requirements.txt
streamlit==1.17.0
tensorflow==2.10.0
Pillow==9.3.0